SystemUI: Fix RemoteDisplayTile crash on clean boot
Previously a patch was put up to fix adding/removing tiles due to a race
condition and not checking with the executor was still active at the
time of onPrepare(), but we did not add the check to unPrepare(). This
bug only seems to surface on a clean boot.

Add the same check to unPrepare() to avoid crashing.

Change-Id: I7485ebed41aac94987305b436777bec45beac762
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/quicksettings/RemoteDisplayTile.java